+++ changelog.md
@@ -1,6 +1,15 @@
 # Version log
 
 ## 0.0.1
+
+### issue#8 rocking curve
+- [x] compute and display rocking curve in side figure
+- [x] add panel for rocking curve parameters
+- [x] show computation evolution
+- [x] go through rocking patterns
+- [x] toggle rocking/overlay images
+- add figure for showing 3d orientation
+
 ### issue#7 (Fri  8 Apr 12:38:20 BST 2022)
 - display b0.beam_vs_thickness in either another window or aside of analysis
 - toggle button for dual view